morse-trainer/requirements.txt
# 基础依赖
kivy==2.2.1
kivymd==1.1.1
numpy==1.26.0
scipy==1.11.3
matplotlib==3.8.0

# AI模型依赖
torch==2.1.0
scikit-learn==1.3.2
tensorboard==2.14.1

# 音频处理
pyaudio==0.2.13

# Android特定
jnius==1.1.0
android==1.0.0

# 图形和UI
kivy-garden.graph==0.4
Pillow==10.0.1

# 工具和优化
multiprocessing==2.6.2.1
asyncio==3.4.3
numba==0.57.1  # 用于加速计算
pyjnius==1.5.0  # Android交互
pytest==7.4.2  # 测试

# 文档和开发
sphinx==7.2.6  # 文档生成
pylint==3.0.1  # 代码质量检查 